Create the defmondrian macro
Simplifies the definition of a mondrian animation, given that all of our
animations use the same boilerplate startup code.  This has (?) to be a
macro because we do not want to pass in a direct reference to the update
pipeline and render stack functions; instead we want those to be
resolved at runtime.